requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.0
pytest==7.4.0
pytest-django==4.5.2
coverage==7.3.2

